Hackathon 3/web-app/requirements.txt
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.1.0
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.9
imageio==2.9.0
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
numpy==1.19.1
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
psycopg2==2.8.5
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
timer==1.0.0
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
update-checker==0.18.0
urllib3==1.25.9
websocket-client==0.57.0
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
